requirements.txt
# ComfyUI Base64Nodes 依赖包
websockets>=11.0.0
aiohttp>=3.8.0
Pillow>=8.0.0
# MiniMind模型依赖
transformers>=4.30.0
torch>=2.0.0
tokenizers>=0.13.0